MSE for the dataset
creates a vandermonte for features and fits polynomial based on SVD approach to find the least squared error. If the degree is greater than the # of knowns then requires a ridge penalty to ensure that vector remains positive definite and invertable. prediction uses resultant beta values and time values to get a prediction of the 4th point.