Updated MSE
This is the evaluation window using a scale for the time because the values on the vandermonte matrix becomes large quickly resulting in potential errors. time scale + ridge produces a better MSE. MSE across data takes the data and calculates mse across all 4 points. Plot capability for actual vs preds for one set.